cmd/trayscale: refactor all UI code into `internal/ui`
