+++ internal/ui/status-icon-active.png
