+++ internal/ui/status-icon-inactive.png
