+++ internal/ui/trayscale.cmb
