+++ internal/ui/ui.go
@@ -1,16 +1,18 @@
-package main
+package ui
 
 import (
 	"context"
-	_ "embed"
+	"io"
 	"log/slog"
 	"net/netip"
 	"os"
 	"slices"
 	"strconv"
+	"strings"
 	"time"
 
 	"deedles.dev/mk"
+	"deedles.dev/trayscale"
 	"deedles.dev/trayscale/internal/tsutil"
 	"deedles.dev/trayscale/internal/version"
 	"deedles.dev/trayscale/internal/xcmp"
@@ -23,6 +25,7 @@ import (
 	"github.com/diamondburned/gotk4/pkg/gtk/v4"
 	"tailscale.com/ipn/ipnstate"
 	"tailscale.com/types/key"
+	"tailscale.com/types/opt"
 )
 
 // App is the main type for the app, containing all of the state
@@ -713,4 +716,88 @@ func (a *App) newPeerPage(status tsutil.Status, peer *ipnstate.PeerStatus) *peer
 	})
 
 	return &page
+}
+
+const (
+	appID                 = "dev.deedles.Trayscale"
+	prefShowWindowAtStart = "showWindowAtStart"
+)
+
+type enum[T any] struct {
+	Index int
+	Val   T
+}
+
+func enumerate[T any](i int, v T) enum[T] {
+	return enum[T]{i, v}
+}
+
+func formatTime(t time.Time) string {
+	if t.IsZero() {
+		return ""
+	}
+	return t.Format(time.StampMilli)
+}
+
+// must returns v if err is nil. If err is not nil, it panics with
+// err's value.
+func must[T any](v T, err error) T {
+	if err != nil {
+		panic(err)
+	}
+	return v
+}
+
+// readAssetString returns the contents of the given embedded asset as
+// a string. It panics if there are any errors.
+func readAssetString(file string) string {
+	var str strings.Builder
+	f := must(trayscale.Assets().Open(file))
+	must(io.Copy(&str, f))
+	return str.String()
+}
+
+func peerName(status tsutil.Status, peer *ipnstate.PeerStatus, self bool) string {
+	const maxNameLength = 30
+	name := tsutil.DNSOrQuoteHostname(status.Status, peer)
+	if len(name) > maxNameLength {
+		name = name[:maxNameLength-3] + "..."
+	}
+
+	if self {
+		return name + " [This machine]"
+	}
+	if peer.ExitNode {
+		return name + " [Exit node]"
+	}
+	if peer.ExitNodeOption {
+		return name + " [Exit node option]"
+	}
+	return name
+}
+
+func peerIcon(peer *ipnstate.PeerStatus) string {
+	if peer.ExitNode {
+		return "network-workgroup-symbolic"
+	}
+	if peer.ExitNodeOption {
+		return "network-server-symbolic"
+	}
+
+	return "folder-remote-symbolic"
+}
+
+func boolIcon(v bool) string {
+	if v {
+		return "emblem-ok-symbolic"
+	}
+	return "window-close-symbolic"
+}
+
+func optBoolIcon(v opt.Bool) string {
+	b, ok := v.Get()
+	if !ok {
+		return "dialog-question-symbolic"
+	}
+	return boolIcon(b)
 }
